Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,11 +2,11 @@
 
 int main() {
     int a,b,c;
-    float x;
+    double x;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
     x = (a+b+c)/3;
-    printf("Average: %.2f",x);
+    printf("Average: %.2lf",x);
     return 0;
 }
